Changes "Reset" Functionality, column count bug fix
Prior to this patch, reset button would just clear the board.
Instead, the reset button now resets back to the last starting configuration
of the board. Additionally, there was a bug present in the algortihm
for counting current blue/green tiles (would double count tiles if they did
not move). Blasted that shit with a can of Raid and called its mother a whore.